+++ src/request/transaction-report.spec.ts
@@ -25,10 +25,37 @@ describe('Device()', () => {
     expect(report).toThrowError('transactionReport.tag');
   });
 
-  it('constructs', () => {
+  it('constructs with ipAddress', () => {
     expect(() => {
       new TransactionReport({
         ipAddress: '1.1.1.1',
+        tag: Tag.CHARGEBACK,
+      });
+    }).not.toThrow();
+  });
+
+  it('constructs with maxmindId', () => {
+    expect(() => {
+      new TransactionReport({
+        maxmindId: '12345678',
+        tag: Tag.CHARGEBACK,
+      });
+    }).not.toThrow();
+  });
+
+  it('constructs with minfraudId', () => {
+    expect(() => {
+      new TransactionReport({
+        minfraudId: '58fa38d8-4b87-458b-a22b-f00eda1aa20',
+        tag: Tag.CHARGEBACK,
+      });
+    }).not.toThrow();
+  });
+
+  it('constructs with transactionId', () => {
+    expect(() => {
+      new TransactionReport({
+        minfraudId: 'abc123',
         tag: Tag.CHARGEBACK,
       });
     }).not.toThrow();