+++ CHANGELOG.md
@@ -1,6 +1,13 @@
 CHANGELOG
 =========
 
+6.2.0
+------------------
+
+* Updated `TransactionReport` to make the `ipAddress` parameter optional. Now
+  the `tag` and at least one of the following paramters must be supplied:
+  `ipAddress`, `maxmindId`, `minfraudID`, `transactionID`.
+
 6.1.0 (2024-04-16)
 ------------------